md: modify dm_io() so it could return bios instead of submitting it
When trying to modify flashcache to request based (current it's bio based), we need
to make request from bios by ourselves, but dm_io() will submit these bios directly,
so we propose to modify the dm_io() to return bios instead of submiting it.

This could also improve the flexibility of dm_io().

Signed-off-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Peng Tao <bergwolf@gmail.com>