md: add new interface 'mk_rq' in target_type
We are now trying to modify flashcache(https://github.com/facebook/flashcache)
to make it request based so that we can let cfq io-controller control the
bandwidth between different io cgroups.

A search in the dm directory tells me that only multipath is a request based dm
target and its functionality is very simple and map_rq() is used to map the
request to different underlying devices.
We can't work in this way because:

1. the request which processed by map_rq() need to be issued to
	different lower devices (disk device and cache device, in flashcache),
	therefore the request can't be totally remapped by simply changing its
	queue and returning DM_MAPIO_REMAPPED in map_rq() like multipath_map()
2. to submit bios drectly in map_rq() (by return DM_MAPIO_SUBMITTED) will cause
	BUG_ON(!irqs_disabled()) in dm_request_fn() because the
	submit_bio()->generic_make_request()->blk_queue_bio() will definitly
	call spin_unlock_irq to enable the irqs

As above,the interface map_rq() provided by devcie-mapper framework is not
enough for an autonomous target, like flashcache.

We propose to add a new mk_rq interface so that we can make the requests by
ourselves.

Signed-off-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Peng Tao <bergwolf@gmail.com>